Cadvisor agent to get metrics for all docker containers:
Previously the cadvisor agent was specific to work only with
CCM project. Now it can get metrics for all docker containers.

+++ cadvisor/getmetrics_cadvisor.py
@@ -39,8 +39,12 @@ newInstanceAvailable = False
 def getindex(colName):
     if colName == "CPU":
         return 1
-    elif colName == "MemUsedB":
+    elif colName == "DiskRead" or colName == "DiskWrite":
         return 2
+    elif colName == "NetworkIn" or colName == "NetworkOut":
+        return 3
+    elif colName == "MemUsed":
+        return 4
 
 def get_ip_address():
     s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
@@ -55,15 +59,8 @@ def update_docker():
     global newInstanceAvailable
     global dockerInstances
 
-    proc = subprocess.Popen(["docker ps --no-trunc | grep -cP 'rubis_apache' | awk '{print $ 1;}'"], stdout=subprocess.PIPE, shell=True)
-    (out, err) = proc.communicate()
-    num_apache = int(out.split("\n")[0])
 
-    proc = subprocess.Popen(["docker ps --no-trunc | grep -cP 'rubis_db' | awk '{print $ 1;}'"], stdout=subprocess.PIPE, shell=True)
-    (out, err) = proc.communicate()
-    num_sql = int(out.split("\n")[0])
-
-    proc = subprocess.Popen(["docker ps --no-trunc | grep -E 'rubis_apache|rubis_db' | awk '{print $ 1;}'"], stdout=subprocess.PIPE, shell=True)
+    proc = subprocess.Popen(["docker ps --no-trunc | awk '{if(NR!=1) print $1}'"], stdout=subprocess.PIPE, shell=True)
     (out, err) = proc.communicate()
     dockers = out.split("\n")
     if os.path.isfile(os.path.join(homepath,datadir+"totalInstances.json")) == False:
@@ -111,14 +108,12 @@ def getmetric():
                     print "unable to get requests from ",cAdvisoraddress
                     sys.exit()
                 continue
-            if num_apache == 0 and num_sql == 0:
-                break
-            if newInstanceAvailable == True:
+            if newInstanceAvailable == True and os.path.isfile(os.path.join(homepath,datadir+date+".csv")) == True:
                 oldFile = os.path.join(homepath,datadir+date+".csv")
                 newFile = os.path.join(homepath,datadir+date+"."+time.strftime("%Y%m%d%H%M%S")+".csv")
                 os.rename(oldFile,newFile)
-            index = len(r.json()["/docker/"+dockers[0]]["stats"])-1
-            time_stamp = r.json()["/docker/"+dockers[0]]["stats"][index]["timestamp"][:19]
+            index = len(r.json()["/system.slice/docker-"+dockers[0]+".scope"]["stats"])-1
+            time_stamp = r.json()["/system.slice/docker-"+dockers[0]+".scope"]["stats"][index]["timestamp"][:19]
             if (time_stamp in counter_time_map.values()):
                 continue
             counter_time_map[counter] = time_stamp
@@ -127,48 +122,47 @@ def getmetric():
             cpu_all = 0
             resource_usage_file = open(os.path.join(homepath,datadir+date+".csv"), 'a+')
             numlines = len(resource_usage_file.readlines())
-            if num_apache == 0 and len(dockers)-1 != len(dockerInstances):
-                log = log + ",NaN"
             for i in range(len(dockers)-1):
                 #get cpu
-                cpu_used = r.json()["/docker/"+dockers[i]]["stats"][index]["cpu"]["usage"]["total"]
-                prev_cpu = r.json()["/docker/"+dockers[i]]["stats"][index-1]["cpu"]["usage"]["total"]
-                cur_cpu = float((cpu_used - prev_cpu)/10000000)
+                index = len(r.json()["/system.slice/docker-"+dockers[i]+".scope"]["stats"])-1
+                cpu_used = r.json()["/system.slice/docker-"+dockers[i]+".scope"]["stats"][index]["cpu"]["usage"]["total"]
+                prev_cpu = r.json()["/system.slice/docker-"+dockers[i]+".scope"]["stats"][index-1]["cpu"]["usage"]["total"]
+                cur_cpu = float(float(cpu_used - prev_cpu)/10000000)
                 cur_cpu = abs(cur_cpu)
                 #get mem
-                curr_mem = r.json()["/docker/"+dockers[i]]["stats"][index]['memory']['usage']
-                mem = float(curr_mem/(1024*1024)) #MB
+                curr_mem = r.json()["/system.slice/docker-"+dockers[i]+".scope"]["stats"][index]['memory']['usage']
+                mem = float(float(curr_mem)/(1024*1024)) #MB
                 mem = abs(mem)
                 #get disk
-                curr_block_num = len(r.json()["/docker/"+dockers[i]]["stats"][index]["diskio"]["io_service_bytes"])
+                curr_block_num = len(r.json()["/system.slice/docker-"+dockers[i]+".scope"]["stats"][index]["diskio"]["io_service_bytes"])
                 curr_io_read = 0
                 curr_io_write = 0
                 prev_io_read = 0
                 prev_io_write = 0
                 for j in range(curr_block_num):
-                    curr_io_read += r.json()["/docker/"+dockers[i]]["stats"][index]["diskio"]["io_service_bytes"][j]["stats"]["Read"]
-                    curr_io_write += r.json()["/docker/"+dockers[i]]["stats"][index]["diskio"]["io_service_bytes"][j]["stats"]["Write"]
-                prev_block_num = len(r.json()["/docker/"+dockers[i]]["stats"][index-1]["diskio"]["io_service_bytes"])
+                    curr_io_read += r.json()["/system.slice/docker-"+dockers[i]+".scope"]["stats"][index]["diskio"]["io_service_bytes"][j]["stats"]["Read"]
+                    curr_io_write += r.json()["/system.slice/docker-"+dockers[i]+".scope"]["stats"][index]["diskio"]["io_service_bytes"][j]["stats"]["Write"]
+                prev_block_num = len(r.json()["/system.slice/docker-"+dockers[i]+".scope"]["stats"][index-1]["diskio"]["io_service_bytes"])
                 prev_io_read = 0
                 prev_io_write = 0
                 for j in range(prev_block_num):
-                    prev_io_read += r.json()["/docker/"+dockers[i]]["stats"][index-1]["diskio"]["io_service_bytes"][j]["stats"]["Read"]
-                    prev_io_write += r.json()["/docker/"+dockers[i]]["stats"][index-1]["diskio"]["io_service_bytes"][j]["stats"]["Write"]
-                io_read = float((curr_io_read - prev_io_read)/(1024*1024)) #MB
-                io_write = float((curr_io_write - prev_io_write)/(1024*1024)) #MB
+                    prev_io_read += r.json()["/system.slice/docker-"+dockers[i]+".scope"]["stats"][index-1]["diskio"]["io_service_bytes"][j]["stats"]["Read"]
+                    prev_io_write += r.json()["/system.slice/docker-"+dockers[i]+".scope"]["stats"][index-1]["diskio"]["io_service_bytes"][j]["stats"]["Write"]
+                io_read = float(float(curr_io_read - prev_io_read)/(1024*1024)) #MB
+                io_write = float(float(curr_io_write - prev_io_write)/(1024*1024)) #MB
                 #get network
-                prev_network_t = r.json()["/docker/"+dockers[i]]["stats"][index-1]["network"]["tx_bytes"]
-                prev_network_r = r.json()["/docker/"+dockers[i]]["stats"][index-1]["network"]["rx_bytes"]
-                curr_network_t = r.json()["/docker/"+dockers[i]]["stats"][index]["network"]["tx_bytes"]
-                curr_network_r = r.json()["/docker/"+dockers[i]]["stats"][index]["network"]["rx_bytes"]
-                network_t = float((curr_network_t - prev_network_t)/(1024*1024)) #MB
-                network_r = float((curr_network_r - prev_network_r)/(1024*1024)) #MB
-                #log = log + "," + str(cur_cpu) + "," + str(io_read) + "," + str(io_write)+ "," + str(network_r)+ "," + str(network_t)+ "," + str(mem)
-                log = log + "," + str(cur_cpu)
+                prev_network_t = r.json()["/system.slice/docker-"+dockers[i]+".scope"]["stats"][index-1]["network"]["tx_bytes"]
+                prev_network_r = r.json()["/system.slice/docker-"+dockers[i]+".scope"]["stats"][index-1]["network"]["rx_bytes"]
+                curr_network_t = r.json()["/system.slice/docker-"+dockers[i]+".scope"]["stats"][index]["network"]["tx_bytes"]
+                curr_network_r = r.json()["/system.slice/docker-"+dockers[i]+".scope"]["stats"][index]["network"]["rx_bytes"]
+                network_t = float(float(curr_network_t - prev_network_t)/(1024*1024)) #MB
+                network_r = float(float(curr_network_r - prev_network_r)/(1024*1024)) #MB
+                log = log + "," + str(cur_cpu) + "," + str(io_read) + "," + str(io_write)+ "," + str(network_r)+ "," + str(network_t)+ "," + str(mem)
+                #log = log + "," + str(cur_cpu)
                 if(numlines < 1):
                     serverType = ["Web", "DB"]
-                    #fields = ["timestamp","CPU","DiskRead","DiskWrite","NetworkIn","NetworkOut","MemUsed"]
-                    fields = ["timestamp","CPU"]
+                    fields = ["timestamp","CPU","DiskRead","DiskWrite","NetworkIn","NetworkOut","MemUsed"]
+                    #fields = ["timestamp","CPU"]
                     if i == 0:
                         fieldnames = fields[0]
                     host = hostname.partition(".")[0]
@@ -177,16 +171,12 @@ def getmetric():
                             continue
                         if(fieldnames != ""):
                             fieldnames = fieldnames + ","
-                        if num_apache == 0:
-                            server = serverType[1]
-                        else:
-                            server = serverType[i]
                         groupid = getindex(fields[k])
-                        metric = fields[k] + "[" + server + "_" + str(ipAddress) + "]"
+                        dockerID = dockers[i]
+                        if len(dockerID) > 12:
+                            dockerID = dockerID[:12]
+                        metric = fields[k] + "[" + dockerID + "_" + host + "]"
                         fieldnames = fieldnames + metric +":"+str(groupid)
-            if num_sql == 0 and len(dockers)-1 != len(dockerInstances):
-                #log = log + ",NaN,NaN,NaN,NaN,NaN,NaN"
-                log = log + ",NaN"
             if(numlines < 1):
                 resource_usage_file.write("%s\n"%(fieldnames))
             print log #is it possible that print too many things?
